move jit into its own isolated space (mostly)
